requirements.txt
altgraph==0.17.4
certifi==2023.11.17
charset-normalizer==3.3.2
idna==3.6
packaging==23.2
pefile==2023.2.7
Pillow==10.1.0
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pywin32-ctypes==0.2.2
requests==2.31.0
urllib3==2.1.0
